split entrypoint for master tool
