+++ src/gateway/tools/master.py
@@ -16,20 +16,14 @@
 Tool to control the master from the command line.
 """
 from __future__ import absolute_import
-from platform_utils import Platform, System
 
-System.import_libs()
-
-import argparse
 import logging
 import shutil
 import sys
-from logging import handlers
-from six.moves.configparser import ConfigParser
+from argparse import Namespace
 
-import constants
-from gateway.initialize import setup_minimal_master_platform
 from ioc import INJECTED, Inject
+from platform_utils import Platform
 from serial_utils import CommunicationTimedOutException
 
 if False:  # MYPY
@@ -40,24 +34,6 @@ if False:  # MYPY
 
 
 logger = logging.getLogger('openmotics')
-
-
-def setup_logger():
-    # type: () -> None
-    """ Setup the OpenMotics logger. """
-
-    logger.setLevel(logging.DEBUG)
-    logger.propagate = False
-
-    handler = logging.StreamHandler()
-    handler.setLevel(logging.INFO)
-    handler.setFormatter(logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s"))
-    logger.addHandler(handler)
-
-    handler = handlers.RotatingFileHandler(constants.get_update_log_location(), maxBytes=3 * 1024 ** 2, backupCount=2)
-    handler.setLevel(logging.DEBUG)
-    handler.setFormatter(logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s"))
-    logger.addHandler(handler)
 
 
 @Inject
@@ -124,45 +100,8 @@ def get_communicator(master_communicator=INJECTED):
     return master_communicator
 
 
-def main():
-    # type: () -> None
-    """ The main function. """
-    parser = argparse.ArgumentParser(description='Tool to control the master.')
-    parser.add_argument('--port', dest='port', action='store_true',
-                        help='get the serial port device')
-    parser.add_argument('--sync', dest='sync', action='store_true',
-                        help='sync the serial port')
-    parser.add_argument('--reset', dest='reset', action='store_true',
-                        help='reset the master')
-    parser.add_argument('--hard-reset', dest='hardreset', action='store_true',
-                        help='perform a hardware reset on the master')
-    parser.add_argument('--version', dest='version', action='store_true',
-                        help='get the version of the master')
-    parser.add_argument('--wipe', dest='wipe', action='store_true',
-                        help='wip the master eeprom')
-    parser.add_argument('--update', dest='update', action='store_true',
-                        help='update the master firmware')
-    parser.add_argument('--master-firmware-classic',
-                        help='path to the hexfile with the classic firmware')
-    parser.add_argument('--master-firmware-core',
-                        help='path to the hexfile with the core+ firmware')
-
-    args = parser.parse_args()
-
-    setup_logger()
-
-    config = ConfigParser()
-    config.read(constants.get_config_file())
-
-    port = config.get('OpenMotics', 'controller_serial')
-    if args.port:
-        print(port)
-        return
-
-    if not any([args.sync, args.version, args.reset, args.hardreset, args.wipe, args.update]):
-        parser.print_help()
-
-    setup_minimal_master_platform(port)
+def master_tool(args):
+    # type: (Namespace) -> None
     platform = Platform.get_platform()
 
     if args.hardreset:
@@ -195,7 +134,3 @@ def main():
             master_factory_reset()
     finally:
         communicator.stop()
-
-
-if __name__ == '__main__':
-    main()
